Add basic interface and pattern rendering
This interface includes a button for loading a file, and a canvas which
renders a preview of the resulting pattern. At the moment the only
pattern is the basic square grid, and there's no mechanism for
downloading the resulting image. A loaded image will be saved in
localStorage, so if the page is refreshed the loaded image will still be
there.