Shrink the preview canvas before resizing
When resizing the preview canvas, it was still in the container used to
calculate the new size of the canvas. This prevented it from ever
getting smaller, because the container is always at least as big as the
canvas inside it. Now the canvas is resized to 0 before the size of the
container is calculated.