Add a save image button to download the preview
Right now the user can't specify the dimensions of the resulting image,
because it just renders the preview canvas. The name and type of the
output depend on the input, which also means that information is also
stored in localStorage.

The event on the file reader has been changed to onload instead of
onloadend, so it only runs if the file successfully loads. An onerror
event should probably also be registered.