Center the preview canvas on corner of the smaller
THere was some extra effort expended in the code to center the preview
canvas on the center of the smaller canvas instead of just the top-left
corner. This seems to mostly produce a worse result thanh centering it
on the corner, so the extra effort was just wasted, and has been
removed.