Add output dimensions, reduce code coupling
The interface now has inputs for the size of the saved image, and the
image is rendered on a separate canvas of that size before being
saved. There is some validation to ensure the dimensions must be
integers and greater than zero, but the save function (or better yet,
the rendering in general, so that preview is included) should probably
include some validation itself.

All of the straightline code has been removed to avoid sharing variables
between different functionality. Instead, each piece of functionality
has been split into its own function, none of which interacts with the
page at all. A much smaller piece of straightline is responsible for
setting up the button events, which includes fetching elements off the
page and passing them to the other functions as a 'rendering' context.