Hide the save dimension inputs until button hover
Only showing the save dimension inputs when the save button is hovered
gets them out of the way until they are needed. The inputs fade out when
the group is no longer hovered over in an attempt to ease the pain of
trying to interact with elements whose appearance on the page are
subject to the position of the user's mouse.

A quick reset of the inputs' values to ensure label clicks put the caret
in the right place is also included.