Scale the original image preview if it is too big
The original image preview could take up too much of the screen if the
image was large, as it rendered it exactly as it was. Now the preview
takes into account the size of the document, and ensures it takes up to
more than a quarter of a dimension of the application.

Comments in the function which rendered the preview have also been
improved, for clarity.