Move state into objects for the main components
Much of the awkward state in the main load event has been offloaded into
objects. This is mostly in preparation for the triangle selector to be
manipulated by the user, as the position of the triangle is now handed
from the selector to the pattern renderer. This way the triangle can use
a single state for its position while being scaled onto the selector and
cats-eye simultaneously.