Bind selector to mouse, clip preview appropriately
As a temporary measure to show the movement of the triangle points, the
leftmost point of the selection triangle is now bound to the x position
of the mouse. The generated pattern is now just rendered from the shape
of the triangle. The preview pattern rendering can run fairly slow to be
updating on every mouse move, so the update of the preview is
constrained by a short timeout which can be cancelled if another mouse
move occurs before the draw begins, which should make the experience
more smooth (at the cost of fully live updating). This should be
revisited: maybe allow the user to customise it (or better yet,
calculate it based on the current performance of the user's browser).