Allow the selection triangle points to be moved
The points of the selection triangle can now be dragged around to
manipulate the area of the image that is clipped for the cats-eye. The
triangle is guaranteed to remain right-angled. As before, the longer
non-hypotenuse edge of the triangle is shortened to the length of the
other, potentially deforming the picture. The over-drawing to correct
missing pixels in some browsers is gone because it relied on a fixed
shape of the triangle, and should be added back in a more general form.

The pattern rendering still assumes that the point opposite the
hypotenuse in the top-right corner. If this is invalidated then the
triangle will continue to render as though it is. This is not the
intended behaviour.

The points of the selection triangle are stored in the local storage, so
all aspects of a selection remain between page loads.