Stop the selection canvas from filling its parent
The background of the selection canvas would extend to fill its parent
if the parent's width exceeded the width of the canvas, which could
happen if the width of the rest of the selection interface exceeded the
width of the current image. Now the canvas background is revealed by
setting its style display property to inline-block instead of block,
meaning it will only take up the necessary amount of room to contain the
selection canvas.