Catch local storage and canvas errors
The application was not very robust in the presence of large images (or
just machines with less memory). Images which exceeded the local storage
limits or the canvas memory limits could raise uncaught errors. These
errors are now caught, reported to the user, and the attempted action is
either cancelled (in the case that the image is too big to even render
on the canvas) or continued (in the case that the image is too big to
store, or no local storage exists at all), with the latter case
including a warning that their changes will not be saved.