Add an option to display the cats-eye grid
The grid reuses the triangle clipping code to stroke two lines along the
triangle edges. Two lines are used to ensure the lines are visible
against any image colour behind them. The toggle is stored in local
storage, like all of the other settings.

The offsets to deal with draw distances in Safari and Firefox are
removed in order to draw the lines properly. A better solution should be
found for that problem.