Reimplement the application in TypeScript
This change reimplements the application in TypeScript, including
splitting the monolithic script file into multiple different modules,
and splitting much of the functionality into different classes. The
result is a much more modular (though still somewhat overly coupled)
codebase. These disparate modules are built back into a (minified)
single JavaScript file by Webpack, first by converting it to ECMAScript
2015 and then running Babel JS to produce code modern browsers can
handle.

The main difference in functionality now is that the triangle selector
is entirely freeform, in preparation for implementing more transform
types. In the future, transforms should be able to indicate the required
shape of the triangle.